Fixed bug in the command prompt in the Aqua version of the ext gui.
When scrolling though command history in the command prompt in the Aqua external gui, the up and down arrow keys were getting confused for inputs. In addition to scrolling through the command history, each press of an arrow key would also dump a garbage (ie not present in most fonts) unicode character into the command prompt. A deep dive into the tkinter arcana revealed a fix, which was to add a `return 'break'` statement to the end of the methods (`back(self)` and `forward(self)`) that get bound to the up and down arrows.